boulder: Normalize Python providers/dependencies properly
Part of aerynos/os-tools#496
